:fire: join `espanso-path` to the main binary crate
