--- espanso-path/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "espanso-path"
-version = "0.1.0"
-authors = ["Federico Terzi <federico-terzi@users.noreply.github.com>"]
-edition = "2021"
-
-[dependencies]
-log.workspace = true
-dirs.workspace = true
-
-[lints]
-workspace = true
