+++ espanso/Cargo.toml
@@ -44,7 +44,6 @@ espanso-match = { path = "../espanso-match" }
 espanso-clipboard = { path = "../espanso-clipboard" }
 espanso-info = { path = "../espanso-info" }
 espanso-render = { path = "../espanso-render" }
-espanso-path = { path = "../espanso-path" }
 espanso-ipc = { path = "../espanso-ipc" }
 espanso-modulo = { path = "../espanso-modulo", optional = true }
 espanso-kvs = { path = "../espanso-kvs" }
