+++ espanso/src/cli/mod.rs
@@ -19,9 +19,9 @@
 
 use std::path::PathBuf;
 
+use crate::path::Paths;
 use clap::ArgMatches;
 use espanso_config::{config::ConfigStore, error::NonFatalErrorSet, matches::store::MatchStore};
-use espanso_path::Paths;
 
 pub mod cmd;
 pub mod daemon;
