+++ espanso/src/cli/package/list.rs
@@ -17,10 +17,10 @@
  * along with espanso.  If not, see <https://www.gnu.org/licenses/>.
  */
 
+use crate::path::Paths;
 use anyhow::{Context, Result};
 use clap::ArgMatches;
 use espanso_package::StoredPackage;
-use espanso_path::Paths;
 
 use crate::info_println;
