+++ espanso/src/cli/package/update.rs
@@ -17,10 +17,10 @@
  * along with espanso.  If not, see <https://www.gnu.org/licenses/>.
  */
 
+use crate::path::Paths;
 use anyhow::{anyhow, Context, Result};
 use clap::ArgMatches;
 use espanso_package::{Archiver, PackageSpecifier, ProviderOptions, SaveOptions, StoredPackage};
-use espanso_path::Paths;
 
 use crate::{error_eprintln, info_println, warn_eprintln};
