+++ espanso/src/cli/worker/engine/process/middleware/image_resolve.rs
@@ -17,7 +17,7 @@
  * along with espanso.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-use espanso_path::Paths;
+use crate::path::Paths;
 
 use espanso_engine::process::PathProvider;
