+++ schemas/matches.json
@@ -1,0 +1,35 @@
+{
+    "$schema": "./match.schema.json",
+    "matches": [
+        {
+            "trigger": "qwe",
+            "replace": "asdf",
+            "vars": [
+                {
+                    "type": "choice",
+                    "name": "auca",
+                    "params": {
+                        "values": [
+                            "adf",
+                            "adf"
+                        ]
+                    }
+                },
+                {
+                    "type": "choice",
+                    "name": "auca2",
+                    "params": {
+                        "values": [
+                            {
+                                "id": "adf",
+                                "label": "d"
+                            },
+                            "adf",
+                            {}
+                        ]
+                    }
+                }
+            ]
+        }
+    ]
+}
